Move self-update into global state
